Modified Screenshot workflow and Added scan option for barcode feature in editor
