+++ lib/l10n/app_en.arb
@@ -22,7 +22,7 @@
   "noImageSelectedFeedback": "Import an image first!",
   "adjustButtonLabel": "Adjust",
   "importImageButtonLabel": "Import",
-  "openEditor": "Editor",
+  "openEditor": "Create",
   "importStartingImageFeedback": "Import an image to begin",
   "transferButtonLabel": "Transfer",
   "filterScreenTitle": "Select a Filter",
