+++ lib/l10n/app_localizations_de.dart
@@ -75,7 +75,7 @@ class AppLocalizationsDe extends AppLocalizations {
   String get importImageButtonLabel => 'Import';
 
   @override
-  String get openEditor => 'Editor';
+  String get openEditor => 'Create';
 
   @override
   String get importStartingImageFeedback => 'Import an image to begin';
