+++ lib/l10n/app_localizations_pt.dart
@@ -75,7 +75,7 @@ class AppLocalizationsPt extends AppLocalizations {
   String get importImageButtonLabel => 'Import';
 
   @override
-  String get openEditor => 'Editor';
+  String get openEditor => 'Create';
 
   @override
   String get importStartingImageFeedback => 'Import an image to begin';
