+++ lib/l10n/app_localizations_vi.dart
@@ -75,7 +75,7 @@ class AppLocalizationsVi extends AppLocalizations {
   String get importImageButtonLabel => 'Import';
 
   @override
-  String get openEditor => 'Editor';
+  String get openEditor => 'Create';
 
   @override
   String get importStartingImageFeedback => 'Import an image to begin';
