+++ lib/pro_image_editor/features/movable_background_image.dart
@@ -8,9 +8,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:barcode_widget/barcode_widget.dart';
 import 'package:magicepaperapp/constants/asset_paths.dart';
 import 'package:magicepaperapp/pro_image_editor/features/bottom_bar.dart';
+import 'package:magicepaperapp/pro_image_editor/features/barcode_editor.dart';
 import 'package:magicepaperapp/pro_image_editor/features/text_bottom_bar.dart';
 import 'package:magicepaperapp/provider/color_palette_provider.dart';
 import 'package:magicepaperapp/provider/getitlocator.dart';
@@ -65,10 +65,6 @@ class _MovableBackgroundImageExampleState
 
   final _bottomTextStyle = const TextStyle(fontSize: 10.0, color: Colors.white);
 
-  final TextEditingController _barcodeController = TextEditingController();
-  Barcode _selectedBarcode = Barcode.qrCode();
-  String _barcodeData = '';
-
   @override
   void initState() {
     super.initState();
@@ -81,18 +77,11 @@ class _MovableBackgroundImageExampleState
     preCacheImage(assetPath: ImageAssets.orangeBoard);
     preCacheImage(assetPath: ImageAssets.yellowBoard);
     _bottomBarScrollCtrl = ScrollController();
-
-    _barcodeController.addListener(() {
-      setState(() {
-        _barcodeData = _barcodeController.text;
-      });
-    });
   }
 
   @override
   void dispose() {
     _bottomBarScrollCtrl.dispose();
-    _barcodeController.dispose();
     super.dispose();
   }
 
@@ -395,297 +384,20 @@ class _MovableBackgroundImageExampleState
     );
   }
 
-  final Map<String, String> barcodeFormatToSupportedChars = {
-    'Aztec': 'All',
-    'CODABAR': '0-9 \$ - . / : +',
-    'CODE 128': 'All',
-    'CODE 39': '0-9 A-Z - . \$ / + % , ',
-    'CODE 93': '0-9 A-Z - . \$ / + % , ',
-    'Data Matrix': 'All',
-    'EAN 13': '0-9',
-    'EAN 2': '0-9',
-    'EAN 5': '0-9',
-    'EAN 8': '0-9',
-    'GS1 128': 'All',
-    'ISBN': '0-9',
-    'ITF': '0-9',
-    'ITF 14': '0-9',
-    'ITF 16': '0-9',
-    'PDF417': 'All',
-    'QR Code': 'All',
-    'RM4SCC': '0-9 A-Z',
-    'Telepen': 'All',
-    'UPC A': '0-9',
-    'UPC E': '0-9',
-  };
-
-  String? _validateBarcodeData(String data, Barcode barcode) {
-    if (data.isEmpty) {
-      return null;
-    }
-    final allowedChars = barcode.charSet.toSet();
-    for (final rune in data.runes) {
-      if (!allowedChars.contains(rune)) {
-        final char = String.fromCharCode(rune);
-        final rules = barcodeFormatToSupportedChars[_selectedBarcode.name];
-        return "Invalid character '$char' \nSupported characters: ${rules ?? 'Please check barcode rules'}";
-      }
-    }
-    if (data.length < barcode.minLength) {
-      return 'Data too short for ${barcode.name}. Minimum length is ${barcode.minLength}.';
-    }
-    if (barcode.maxLength < 10000 && data.length > barcode.maxLength) {
-      return 'Data too long for ${barcode.name}. Maximum length is ${barcode.maxLength}.';
-    }
-    return null;
-  }
-
   void _openBarcodeSheet() {
     showModalBottomSheet(
       context: context,
       isScrollControlled: true,
       builder: (context) {
-        return StatefulBuilder(
-          builder: (context, setModalState) {
-            return Container(
-              height: MediaQuery.of(context).size.height * 0.75,
-              padding: const EdgeInsets.all(16.0),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                children: [
-                  const Text(
-                    'Add Barcode',
-                    style: TextStyle(fontSize: 24, fontWeight: FontWeight.w500),
-                    textAlign: TextAlign.center,
-                  ),
-                  const SizedBox(height: 20),
-                  TextField(
-                    controller: _barcodeController,
-                    decoration: const InputDecoration(
-                      labelText: 'Barcode Data',
-                      hintText: 'Enter barcode data',
-                      prefixIcon: Icon(Icons.qr_code_2_rounded),
-                      border: OutlineInputBorder(),
-                    ),
-                    onChanged: (value) {
-                      setModalState(() {
-                        _barcodeData = value;
-                      });
-                    },
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.only(top: 8.0, left: 4.0),
-                    child: Text(
-                      'Characters: ${_barcodeData.length}',
-                      style: TextStyle(
-                        color: Colors.grey[600],
-                        fontSize: 12,
-                        fontWeight: FontWeight.w500,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(height: 20),
-                  _buildBarcodeFormatSelector(setModalState),
-                  const SizedBox(height: 20),
-                  Expanded(
-                    child: Center(
-                      child: _buildBarcodePreviewWidget(),
-                    ),
-                  ),
-                  const SizedBox(height: 20),
-                  if (_barcodeData.isNotEmpty &&
-                      _validateBarcodeData(_barcodeData, _selectedBarcode) ==
-                          null)
-                    ElevatedButton(
-                      onPressed: () {
-                        _addBarcodeLayer();
-                        Navigator.pop(context);
-                      },
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.green,
-                        foregroundColor: Colors.white,
-                        padding: const EdgeInsets.symmetric(vertical: 12),
-                      ),
-                      child: const Text('Add Barcode'),
-                    ),
-                ],
-              ),
-            );
+        return BarcodeEditor(
+          onBarcodeCreated: (layer) {
+            editorKey.currentState?.addLayer(layer);
+            Navigator.pop(context);
           },
+          initialScale: _initScale,
         );
       },
     );
-  }
-
-  Widget _buildBarcodeFormatSelector(StateSetter setModalState) {
-    final Map<String, Barcode> availableFormats = {
-      'QR Code': Barcode.qrCode(),
-      'Data Matrix': Barcode.dataMatrix(),
-      'Aztec': Barcode.aztec(),
-      'PDF417': Barcode.pdf417(),
-      'Code 128': Barcode.code128(),
-      'Code 93': Barcode.code93(),
-      'Code 39': Barcode.code39(),
-      'Codabar': Barcode.codabar(),
-      'EAN-13': Barcode.ean13(),
-      'EAN-8': Barcode.ean8(),
-      'EAN-5': Barcode.ean5(),
-      'EAN-2': Barcode.ean2(),
-      'GS1 128': Barcode.gs128(),
-      'ISBN': Barcode.isbn(),
-      'ITF': Barcode.itf(),
-      'ITF-16': Barcode.itf16(),
-      'ITF-14': Barcode.itf14(),
-      'RM4SCC': Barcode.rm4scc(),
-      'Telepen': Barcode.telepen(),
-      'UPC-A': Barcode.upcA(),
-      'UPC-E': Barcode.upcE(),
-    };
-
-    return DropdownButtonFormField<String>(
-      value: _selectedBarcode.name,
-      decoration: const InputDecoration(
-        labelText: 'Barcode Format',
-        border: OutlineInputBorder(),
-      ),
-      items: availableFormats.entries
-          .map((entry) => DropdownMenuItem(
-                value: entry.value.name,
-                child: Text(entry.key),
-              ))
-          .toList(),
-      onChanged: (newBarcodeName) {
-        if (newBarcodeName != null) {
-          setModalState(() {
-            _selectedBarcode = availableFormats.values.firstWhere(
-              (barcode) => barcode.name == newBarcodeName,
-              orElse: () => Barcode.qrCode(),
-            );
-          });
-        }
-      },
-    );
-  }
-
-  Widget _buildBarcodePreviewWidget() {
-    if (_barcodeData.isEmpty) {
-      return Container(
-        width: 240,
-        height: 120,
-        decoration: BoxDecoration(
-          color: Colors.grey[200],
-          border: Border.all(color: Colors.grey),
-          borderRadius: BorderRadius.circular(8),
-        ),
-        child: const Center(
-          child: Text(
-            'Enter barcode data to see preview',
-            style: TextStyle(color: Colors.grey),
-          ),
-        ),
-      );
-    }
-
-    // Check for validation errors before trying to build the barcode
-    final validationError =
-        _validateBarcodeData(_barcodeData, _selectedBarcode);
-    if (validationError != null) {
-      return Container(
-        width: double.infinity,
-        height: 200,
-        decoration: BoxDecoration(
-          color: Colors.red[50],
-          border: Border.all(color: Colors.red[400]!, width: 2),
-          borderRadius: BorderRadius.circular(12),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(12.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Icon(
-                Icons.error_outline,
-                color: Colors.red[600],
-                size: 45,
-              ),
-              const SizedBox(height: 8),
-              Text(
-                'Invalid Barcode',
-                style: TextStyle(
-                  color: Colors.red[700],
-                  fontSize: 20,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              const SizedBox(height: 4),
-              Expanded(
-                child: SingleChildScrollView(
-                  child: Text(
-                    validationError,
-                    style: TextStyle(
-                      color: Colors.red[600],
-                      fontSize: 16,
-                      fontWeight: FontWeight.w500,
-                    ),
-                    textAlign: TextAlign.center,
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      );
-    }
-
-    return BarcodeWidget(
-      style: const TextStyle(color: Colors.black),
-      padding: const EdgeInsets.all(10),
-      backgroundColor: Colors.white,
-      barcode: _selectedBarcode,
-      data: _barcodeData,
-    );
-  }
-
-  void _addBarcodeLayer() {
-    final validationError =
-        _validateBarcodeData(_barcodeData, _selectedBarcode);
-    if (validationError != null) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: Text(validationError),
-          backgroundColor: Colors.red,
-        ),
-      );
-      return;
-    }
-
-    final barcodeWidget = BarcodeWidget(
-      barcode: _selectedBarcode,
-      data: _barcodeData,
-      style: const TextStyle(color: Colors.black),
-      backgroundColor: Colors.white,
-      padding: const EdgeInsets.all(5),
-    );
-
-    editorKey.currentState?.addLayer(
-      WidgetLayer(
-        offset: Offset.zero,
-        scale: _initScale * 1.25,
-        widget: Container(
-          child: barcodeWidget,
-        ),
-        interaction: LayerInteraction(
-          enableEdit: true,
-          enableMove: true,
-          enableRotate: true,
-          enableScale: true,
-          enableSelection: true,
-        ),
-      ),
-    );
-
-    _barcodeController.clear();
-    _barcodeData = '';
   }
 
   Future<Uint8List> _loadAndResizeWhiteBoard() async {
